+++ src/main/java/nextstep/auth/application/dto/GithubAccessTokenResponse.java
@@ -1,4 +1,4 @@
-package nextstep.favorite.application.dto;
+package nextstep.auth.application.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
